refactor(front): Extract `ImagePreview` component
