+++ src/Pages/Photo.module.css
@@ -1,8 +1,7 @@
 .grid {
   &.hidden img,
   &.hidden a.gridItem,
-  img.hidden,
-  div.loading.hidden {
+  img.hidden {
     opacity: 0;
   }
 }
@@ -27,29 +26,6 @@
     padding-left: 0;
   }
 
-  > img {
-    max-width: 100%;
-    height: auto;
-    display: block;
-    transition: opacity 0.3s ease-in-out;
-    border-radius: 1rem;
-  }
-
-  > div.loading {
-    position: absolute;
-    display: flex;
-    align-items: center;
-    height: 100%;
-    width: 100%;
-    transition: opacity 0.3s ease-in-out;
-
-    > svg {
-      display: block;
-      margin: 0 auto;
-      z-index: 10;
-    }
-  }
-
   &:hover {
     /* .gridItem is an <a>, unset highlighting on hover */
     color: white;
@@ -65,6 +41,7 @@
 
   .fragmentIdentifier {
     position: absolute;
+    top: 0;
     display: flex;
     align-items: center;
     height: 100%;
